Backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==1.10.7
requests==2.31.0
aiofiles==23.2.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Pillow==11.2.1